Make WriteToFile pipeline more robust
Currently the pipeline will try to write given items to file but
if the items do not implement String.Chars it will crash. By using
Kernel.inspect any term can be saved without having to explicitly
implement the protocol.

This was discovered in an example where the pipeline tried to write
Crawly.Request structs but failed due to the struct not implementing
String.Chars.