Spider and Configuration generators
The current commit introduces two mix commands which are supposed to speed
up spider/configuration creation for Crawly projects.

This is one of the features used in Scrapy, and it allows you to make spiders
almost instantly withjust one command. It simplifies my work, and hopefully, it will be helpful for others as well!

Mix files are not included into coverage reports