Switch Crawly to use GH actions (#289)
I decided to use Gh actions instead of circle/coveralls,
as I am getting quite a bit of problems when PRs are not
adequately built, or tests fail because overalls do not
accept the new report.

The new version is quite simplistic.
It does not depend on external tools (just github). And I hope builds will be more stable.