Fix Crawly.Utils.list_spiders/0
The list_spiders function is a bit tricky, it was using __info__ method,
that as I see does not work in quite a few cases. Even worse it was
just hanging when :shell_defaults.__info__ was called.

I decided to change the function to use .module_info instead, and added
some debug loggings